`Integration Tests`: split up tests in smaller files
With #2069 this file grew up a bit too much.
This splits up `BaseStoreKitIntegrationTests` and `StoreKitObserverModeIntegrationTests` to make them more maintanable.
